pyproject.toml
[project]
name = "auto-emailer"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "ddgs>=9.0.0",
    "apscheduler>=3.10.0",
    "trafilatura>=2.0.0",
    "requests>=2.28.0",
]

[project.scripts]
auto-emailer = "auto_emailer.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
